requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
aiohttp==3.8.1
    # via -r requirements.in
aiosignal==1.2.0
    # via aiohttp
aniso8601==9.0.1
    # via flask-restful
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via aiohttp
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via flask
flask==2.1.3
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-restful
    #   flask-sqlalchemy
    #   slackeventsapi
flask-cors==3.0.10
    # via -r requirements.in
flask-restful==0.3.9
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via -r requirements.in
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.3
    # via
    #   requests
    #   yarl
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
pyee==9.0.4
    # via slackeventsapi
python-dotenv==0.20.0
    # via -r requirements.in
pytz==2022.1
    # via flask-restful
requests==2.28.1
    # via -r requirements.in
six==1.16.0
    # via
    #   flask-cors
    #   flask-restful
slack==0.0.2
    # via -r requirements.in
slack-sdk==3.18.1
    # via -r requirements.in
slackeventsapi==3.0.1
    # via -r requirements.in
sqlalchemy==1.4.39
    # via flask-sqlalchemy
typing-extensions==4.3.0
    # via pyee
urllib3==1.26.11
    # via requests
werkzeug==2.2.1
    # via flask
yarl==1.7.2
    # via aiohttp
